+++ src/kafunk/Protocol.fs
@@ -365,6 +365,12 @@ module Protocol =
     let [<Literal>] CREATE_TIME = 0
     let [<Literal>] LOG_APPEND_TIME = 1
 
+  module internal Record =
+    
+    /// The minimum bytes overhead for a record in a record batch.
+    /// If less than this value of space is available in the buffer, the record is considered partial and skipped.
+    let MIN_RECORD_OVERHEAD = 1 + 1 + 1 + 1 + 1 + 1 + 1 // length attrs timestamp offset key value headers
+
   module internal RecordBatch =
 
     let OFFSET_OFFSET = 0
@@ -427,7 +433,8 @@ module Protocol =
       Binary.sizeBytes m.value
 
     /// Returns the size of the record, excluding the size of the length field itself.
-    static member SizeRecord timestampDelta offsetDelta (m:Message) =
+    /// NB: vartint encodings are used on timestamp delta and offset delta
+    static member SizeRecord (timestampDelta:int64) (offsetDelta:int) (m:Message) =
       let size =
         1 + // attributes
         Binary.sizeVarint64 timestampDelta +
@@ -435,7 +442,7 @@ module Protocol =
         Binary.sizeBytesVarint m.key +
         Binary.sizeBytesVarint m.value +
         Binary.sizeVarint 0 // headers length
-      size //+ Binary.sizeVarint size
+      size
 
     static member Write (m:Message, buf:BinaryZipper) =      
       buf.ShiftOffset 4 // CRC
@@ -447,12 +454,11 @@ module Protocol =
       buf.WriteBytes m.key
       buf.WriteBytes m.value
       let crc = Crc.crc32 buf.Buffer.Array offsetAfterCrc (buf.Buffer.Offset - offsetAfterCrc)
-      //Binary.pokeInt32In (int crc) buf.Buffer.Array (offsetAfterCrc - 4)
-      Binary.pokeUInt32In (int64 crc) buf.Buffer.Array (offsetAfterCrc - 4)
+      Binary.pokeInt32In (int crc) buf.Buffer.Array (offsetAfterCrc - 4)
+      //Binary.pokeUInt32In (int64 crc) buf.Buffer.Array (offsetAfterCrc - 4)
 
     static member Read (magicByte:int8, buf:BinaryZipper) =
       let crc = buf.ReadInt32 ()
-      //let _magicByte = buf.ReadInt8 ()
       buf.ShiftOffset 1 // magic byte
       let attrs = buf.ReadInt8 ()
       let timestamp = 
@@ -465,7 +471,7 @@ module Protocol =
       Message(crc,magicByte,attrs,timestamp,key,value)
 
     // NB: assumes that m.key and m.value use the same underlying array
-    static member ComputeCrc (magicByte:int8, m:Message) =
+    static member ComputeCrc (m:Message) =
       let offsetAtKey =
         m.value.Offset
         - 4 // key length
@@ -475,15 +481,15 @@ module Protocol =
         offsetAtKey
         - 1 // magicByte
         - 1 // attrs
-        - (if magicByte >= 1y then 8 else 0) // timestamp
+        - (if m.magicByte >= 1y then 8 else 0) // timestamp
       let offsetAtEnd = m.value.Offset + m.value.Count
       let readMessageSize = offsetAtEnd - offsetAfterCrc
       let crc32 = Crc.crc32 m.value.Array offsetAfterCrc readMessageSize
       int32 crc32
     
     // NB: assumes that m.key and m.value use the same underlying array
-    static member CheckCrc (magicByte:int8, m:Message) =
-      let crc' = Message.ComputeCrc (magicByte,m)
+    static member CheckCrc (m:Message) =
+      let crc' = Message.ComputeCrc m
       if m.crc <> crc' then
         raise (CorruptCrc32Exception(sprintf "Corrupt message data. Computed CRC32=%i received CRC32=%i|key=%s" crc' m.crc (Binary.toString m.key)))
 
@@ -507,13 +513,14 @@ module Protocol =
     end
   with
 
+    /// Computes the message set size assuming magicByte <= 1y
     static member Size (x:MessageSet) =
       let mutable size = 0
       for i = 0 to x.messages.Length - 1 do
         size <- size + 
           8 + // offset
           4 + // message size field
-          x.messages.[i].messageSize // (Message.Size m)
+          x.messages.[i].messageSize
       size
 
     static member Write (ms:MessageSet, buf:BinaryZipper) =
@@ -574,11 +581,10 @@ module Protocol =
       size
 
     static member WriteRecordBatch (batchLength:int, ms:MessageSet, buf:BinaryZipper) =
-      let firstOffset = 0L //ms.messages.[0].offset
+      let firstOffset = 0L
       let firstTimestamp = ms.messages.[0].message.timestamp
       let maxTimestamp = ms.messages.[ms.messages.Length - 1].message.timestamp
-      //let lastOffset = ms.messages.[ms.messages.Length - 1].offset
-      buf.WriteInt64 firstOffset
+      buf.WriteInt64 firstOffset // first offset
       buf.WriteInt32 (batchLength - RecordBatch.LOG_OVERHEAD)
       buf.WriteInt32 0 // partition leader epoch
       buf.WriteInt8 2y // magic v2
@@ -586,16 +592,16 @@ module Protocol =
       buf.ShiftOffset 4 // crc32 (written later)
       let attributesOffset = buf.Buffer.Offset
       buf.WriteInt16 0s // ((0s <<< 3) ||| 0s) // attributes (TODO)
-      buf.WriteInt32 (ms.messages.Length - 1) //(int (lastOffset - firstOffset)) // last offset delta
+      buf.WriteInt32 (ms.messages.Length - 1) // last offset delta
       buf.WriteInt64 firstTimestamp // first timestamp
       buf.WriteInt64 maxTimestamp // max timestamp
       buf.WriteInt64 -1L // producer id
-      buf.WriteInt16 0s // producer epoch
-      buf.WriteInt32 -1  // first sequence
+      buf.WriteInt16 -1s //0s // producer epoch
+      buf.WriteInt32 -1 // -1  // first sequence
       buf.WriteInt32 ms.messages.Length // num records
       for i = 0 to ms.messages.Length - 1 do
         let msi = ms.messages.[i]
-        let offsetDelta = int (msi.offset - firstOffset)
+        let offsetDelta = i
         let timestampDelta = msi.message.timestamp - firstTimestamp
         buf.WriteVarint msi.messageSize // length
         buf.WriteInt8 0y // attributes
@@ -612,7 +618,6 @@ module Protocol =
       let mss = ResizeArray<MessageSetItem>()
       let mutable lastOffset = 0L
       while buf.Buffer.Count > RecordBatch.RECORD_BATCH_OVERHEAD do
-        //printfn "reading record batch|buffer=%i" buf.Buffer.Count
         let lo = MessageSet.ReadFromRecordBatch (partition,ec,messageSetSize,buf,mss)
         lastOffset <- max lastOffset lo
       MessageSet(mss.ToArray(),lastOffset)
@@ -624,31 +629,34 @@ module Protocol =
       if sizeInBytes < RecordBatch.RECORD_BATCH_OVERHEAD then
         failwithf "record_batch_corrupt_size|first_offset=%i size=%i minimum_allowed_size=%i" firstOffset sizeInBytes RecordBatch.RECORD_BATCH_OVERHEAD
       if recordBatchLength + RecordBatch.LOG_OVERHEAD > messageSetSize then        
-        let errMsg = sprintf "p=%i ec=%i o=%i mss=%i" partition ec firstOffset messageSetSize
+        let errMsg = sprintf "p=%i ec=%i o=%i mss=%i record_batch_length=%i" partition ec firstOffset messageSetSize recordBatchLength
         raise (MessageTooBigException(errMsg, 2y, messageSetSize)) else
       let _partitionLeaderEpoch = buf.ReadInt32 ()
       let magicByte = buf.ReadInt8()
-      let _crcSum = buf.ReadInt32()
+      let crcSum = buf.ReadInt32()
+      let attributesOffset = buf.Buffer.Count
       let batchAttributes = buf.ReadInt16()
       let compressionType = batchAttributes &&& int16 CompressionCodec.Mask // TODO: handle compression
+      if compressionType <> 0s then failwith "compression not supported" else
       let timestampType = 
         if (batchAttributes &&& int16 TimestampType.MASK) = 0s then TimestampType.CREATE_TIME 
         else TimestampType.LOG_APPEND_TIME
       let isControl = batchAttributes &&& int16 0x20 > 0s
       let isTx = batchAttributes &&& int16 0x10 > 0s
       let lastOffsetDelta = buf.ReadInt32()
-      let lastOffset = firstOffset + int64 lastOffsetDelta      
+      let lastOffset = firstOffset + int64 lastOffsetDelta
       let firstTimestamp = buf.ReadInt64()
       let maxTimestamp = buf.ReadInt64()
       let _producerId = buf.ReadInt64()
       let _producerEpoch = buf.ReadInt16()
       let _firstSequence = buf.ReadInt32()
       let numRecords = buf.ReadInt32() // Note the response isn't guaranteed to contain every record in this count!
-      if numRecords < 0 then failwithf "invalid record count=%i" numRecords
+      if numRecords < 0 then failwithf "invalid record count=%i" numRecords else
       //let mutable consumed = RecordBatch.RECORD_BATCH_OVERHEAD //buf.Buffer.Offset - recordBatchStart
       //printfn "p=%i _recordBatchLength=%i mss=%i _numRecords=%i last_offset=%i batch_attrs=%i consumed=%i magic=%i tst=%i payload=%i" partition recordBatchLength messageSetSize numRecords lastOffset batchAttributes consumed magicByte timestampType payloadSize
       let mutable readCount = 0
-      while numRecords > readCount && buf.Buffer.Count > 0 do
+      let minRecordOverhead = Record.MIN_RECORD_OVERHEAD
+      while numRecords > readCount && buf.Buffer.Count > minRecordOverhead do
         let recordLength = buf.ReadVarint()
         let recordStart = buf.Buffer.Offset
         if buf.Buffer.Count < recordLength then
@@ -682,15 +690,21 @@ module Protocol =
         with ex ->
           let msg = sprintf "num_records=%i read_records=%i buffer=%i record_length=%i" numRecords readCount buf.Buffer.Count recordLength
           raise (Exception(msg, ex))
+      
+      //TODO: run check elsewhere?
+      //let crcCount = buf.Buffer.Count - attributesOffset
+      //let crc = Crc.crc32C buf.Buffer.Array attributesOffset crcCount
+      //if int crc <> crcSum then raise (CorruptCrc32Exception ("")) else
+
       //if numRecords = arr.Count then 
       //  printfn "matching_record_counts|buffer=%i" buf.Buffer.Count
       //printfn "p=%i consumed=%i _recordBatchLength=%i recordBatchSize=%i _numRecords=%i last_offset=%i batch_attrs=%i read_count=%i ct=%i" 
       //  partition consumed recordBatchLength messageSetSize numRecords lastOffset batchAttributes readCount compressionType
       lastOffset
 
-    static member CheckCrc (magicByte:int8, ms:MessageSet) =
+    static member CheckCrc (ms:MessageSet) =
       for x in ms.messages do
-        Message.CheckCrc (magicByte,x.message)
+        Message.CheckCrc (x.message)
 
 
   /// A Kafka broker consists of a node id, host name and TCP port.
@@ -999,10 +1013,26 @@ module Protocol =
           buf.WriteInt32 maxBytes
 
   [<NoEquality;NoComparison>]
+  type FetchResponsePartitionItem =
+    struct
+      val partition : Partition
+      val errorCode : ErrorCode
+      val highWatermarkOffset : HighwaterMarkOffset
+      val lastStableOffset : LastStableOffset
+      val logStartOffset : LogStartOffset
+      val abortedTxns : AbortedTransaction[]
+      val messageSetSize : MessageSetSize
+      val messageSet : MessageSet
+      new (p,ec,hwo,lastSo,logSo,txns,mss,ms) =
+        { partition = p ; errorCode = ec ; highWatermarkOffset = hwo ; lastStableOffset = lastSo ; 
+          logStartOffset = logSo ; abortedTxns = txns ; messageSetSize = mss ; messageSet = ms }
+    end
+
+  [<NoEquality;NoComparison>]
   type FetchResponse =
     struct
       val throttleTime : ThrottleTime
-      val topics : (TopicName * (Partition * ErrorCode * HighwaterMarkOffset * LastStableOffset * LogStartOffset * AbortedTransaction[] * MessageSetSize * MessageSet)[])[]
+      val topics : (TopicName * FetchResponsePartitionItem[])[]
       new (tt,topics) = { throttleTime = tt ; topics = topics }
     end
   with
@@ -1022,37 +1052,31 @@ module Protocol =
           let partition = buf.ReadInt32 ()
           let errorCode = buf.ReadInt16 ()
           let highWatermark = buf.ReadInt64 ()
-          let _lastStableOffset = if ver >= 4s then buf.ReadInt64() else -1L
-          let _logStartOffset = if ver >= 5s then buf.ReadInt64() else -1L
-          let _numAbortedTxns = if ver >= 4s then buf.ReadInt32() else -1
-          let _abortedTxns =
-            if _numAbortedTxns >= 0 then
-              let _abortedTxns = Array.zeroCreate _numAbortedTxns
-              for k = 0 to _abortedTxns.Length - 1 do
+          let lastStableOffset = if ver >= 4s then buf.ReadInt64() else -1L
+          let logStartOffset = if ver >= 5s then buf.ReadInt64() else -1L
+          let numAbortedTxns = if ver >= 4s then buf.ReadInt32() else -1
+          let abortedTxns =
+            if numAbortedTxns >= 0 then
+              let abortedTxns = Array.zeroCreate numAbortedTxns
+              for k = 0 to abortedTxns.Length - 1 do
                 let producerId = buf.ReadInt64 ()
                 let firstOffset = buf.ReadInt64 ()
-                _abortedTxns.[k] <- (producerId, firstOffset)
-              _abortedTxns
+                abortedTxns.[k] <- (producerId, firstOffset)
+              abortedTxns
             else
               null
           let messageSetSize = buf.ReadInt32 ()
           if messageSetSize < 0 then failwithf "corrupt_fetch_response_partition_size|size=%i buffer=%i" messageSetSize buf.Buffer.Count
-          let innerBuf = 
-            let buf = ArraySegment(buf.Buffer.Array, buf.Buffer.Offset, messageSetSize)
-            BinaryZipper(buf)
-          let messageSet = MessageSet.Read (MessageVersions.fetchResMessage ver,partition,errorCode,messageSetSize,false,buf)
-            //if messageSetSize = 0 then MessageSet([||]) else
-            //let magicByte = innerBuf.TryPeekIn8AtOffset 16
-            //if magicByte >= 2y || ver >= 3s then 
-            //  let ms = MessageSet.ReadFromRecordBatches (partition,errorCode,messageSetSize,innerBuf)
-            //  buf.ShiftOffset messageSetSize // TODO: remove this?
-            //  ms
-            //else 
-            //  MessageSet.Read (MessageVersions.fetchResMessage ver,partition,errorCode,messageSetSize,false,buf)
-              //MessageSet.Read (MessageVersions.fetchResMessage ver,partition,errorCode,messageSetSize,false,innerBuf)
-              //MessageSet.Read (magicByte,partition,errorCode,messageSetSize,false,innerBuf)
-          partitions.[j] <- partition, errorCode, highWatermark, -1L, -1L, null, messageSetSize, messageSet
-          //buf.ShiftOffset messageSetSize // TODO: remove this?
+          let innerBuf = buf.Limit messageSetSize
+          let messageSet =
+            if messageSetSize = 0 then MessageSet([||]) else
+            let magicByte = innerBuf.TryPeekIn8AtOffset 16
+            if magicByte >= 2y || ver >= 3s then 
+              MessageSet.ReadFromRecordBatches (partition,errorCode,messageSetSize,innerBuf)
+            else 
+              MessageSet.Read (magicByte,partition,errorCode,messageSetSize,false,innerBuf)
+          buf.ShiftOffset messageSetSize
+          partitions.[j] <- FetchResponsePartitionItem(partition, errorCode, highWatermark, lastStableOffset, logStartOffset, abortedTxns, messageSetSize, messageSet)        
         topics.[i] <- (topic,partitions)
       FetchResponse(throttleTime, topics)