+++ RELEASE_NOTES.md
@@ -1,3 +1,8 @@
+### 0.1.15-alpha02 - 14.3.2018
+* IMPROVE: support for v0.11 of broker protocol, including v5 of Fetch and v3 of Produce APIs
+* BREAKING: FetchResponse item now explicit structure rather than tuple. Breaking only if using low-level Fetch API.
+* FEATURE: ProducerQueueType to allow configuration of producer message queue networks.
+
 ### 0.1.15-alpha01 - 5.3.2018
 * BUG: consumer would enter an infinite loop between `offsets_out_of_range` and `resuming_fetch_from_reset_offsets`
 * BUG: stalled consumers after TCP connection timeout